+++ pkg/jq/execution/errors.go
@@ -1,4 +1,4 @@
-package runner
+package execution
 
 import "fmt"
