+++ pkg/jq/execution/runner_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=interface.go -destination=runner/runner_mock.go -package=runner Runner
+//	mockgen -source=interface.go -destination=runner_mock.go -package=execution Runner
 //
 
-// Package runner is a generated GoMock package.
-package runner
+// Package execution is a generated GoMock package.
+package execution
 
 import (
 	context "context"
